kafka_spark/requirements.txt
# python 3.9
pandas
numpy
seaborn
ujson
spotipy
urllib3==1.25.1
pyspark
kafka-python